backy/pyproject.toml
[tool.poetry]
name = "backy"
version = "0.1.0"
description = ""
authors = ["Gerson Gallo <ggallovalle@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.22"
Flask-Migrate = "^3.0.1"
Flask-SQLAlchemy = "^2.5.1"
Flask = "^2.0.1"
Flask-GraphQL = "^2.0.1"
graphene = "^2.1.9"
python-dotenv = "^0.19.0"
toolz = "^0.11.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
